Handle special IWD journal quest entry
IWD has a special "Important Events" quest, that seems to be always the first
one and contains both active and completed entries, which ones are shown is
determined by the journalMode variable value in the entryEnabled function.

This commit adds a compatibility function that splits the quest in two,
each containing either active or non-active entries.

Fixes #13